Move info_cache() and info_datasets() methods from api.py to info.py
