+++ dbcollection/core/api.py
@@ -26,65 +26,4 @@ from .add import add
 from .remove import remove
 from .config_cache import config_cache
 from .query import query
-from .info import InfoCacheAPI, InfoDatasetAPI
-
-
-def info_cache(name=None, paths_info=True, datasets_info=True, categories_info=True,
-               verbose=True, is_test=False):
-    """Prints the cache contents and other information.
-
-    Parameters
-    ----------
-    name : str/list/tuple, optional
-        Name or list of names of datasets to be selected for print.
-    paths_info : bool, optional
-        Print the paths info to screen.
-    datasets_info : bool, optional
-        Print the datasets info to screen.
-    categories_info : bool, optional
-        Print the categories keywords info to screen.
-    verbose : bool, optional
-        Displays text information (if true).
-    is_test : bool, optional
-        Flag used for tests.
-
-    Raises
-    ------
-    TypeError
-        If input arg name is not a string or list/tuple.
-    """
-    printer = InfoCacheAPI(name=name,
-                           paths_info=paths_info,
-                           datasets_info=datasets_info,
-                           categories_info=categories_info,
-                           verbose=verbose,
-                           is_test=is_test)
-
-    printer.run()
-
-
-def info_datasets(db_pattern='', show_downloaded=True, show_available=True,
-                  verbose=True, is_test=False):
-    """Prints information about available and downloaded datasets.
-
-    Parameters
-    ----------
-    db_pattern : str
-        String for matching dataset names available for downloading in the database.
-    show_downloaded : bool, optional
-        Print the downloaded datasets stored in cache.
-    show_available : bool, optional
-        Print the available datasets for load/download with dbcollection.
-    verbose : bool, optional
-        Displays text information (if true).
-    is_test : bool, optional
-        Flag used for tests.
-
-    """
-    printer = InfoDatasetAPI(db_pattern=db_pattern,
-                             show_downloaded=show_downloaded,
-                             show_available=show_available,
-                             verbose=verbose,
-                             is_test=is_test)
-
-    printer.run()
+from .info import info_cache, info_datasets
